fix: extract expression generation to avoid nanoid in workflow bundle
Move visualConditionToExpression and its pure helpers into
condition-expression.ts so condition-resolver.ts (used by the workflow
executor) no longer transitively imports nanoid, which the workflow
runtime rejects as a Node.js module.
