+++ keeperhub/lib/condition-builder-utils.ts
@@ -58,135 +58,7 @@ export function createEmptyGroup(): ConditionGroup {
   };
 }
 
-function wrapOperand(operand: string): string {
-  const trimmed = operand.trim();
-  if (!trimmed) {
-    return '""';
-  }
-
-  // Template references like {{@nodeId:Label.field}} pass through as-is
-  // The runtime template resolver handles these before eval
-  if (trimmed.startsWith("{{") && trimmed.endsWith("}}")) {
-    return trimmed;
-  }
-
-  // Numeric values pass through
-  if (!Number.isNaN(Number(trimmed))) {
-    return trimmed;
-  }
-
-  // Boolean literals
-  if (trimmed === "true" || trimmed === "false") {
-    return trimmed;
-  }
-
-  // null / undefined
-  if (trimmed === "null" || trimmed === "undefined") {
-    return trimmed;
-  }
-
-  // Already quoted strings pass through
-  if (
-    (trimmed.startsWith('"') && trimmed.endsWith('"')) ||
-    (trimmed.startsWith("'") && trimmed.endsWith("'"))
-  ) {
-    return trimmed;
-  }
-
-  // Default: treat as string literal and quote it
-  const escaped = trimmed.replace(/\\/g, "\\\\").replace(/"/g, '\\"');
-  return `"${escaped}"`;
-}
-
-function ruleToExpression(rule: ConditionRule): string {
-  const left = wrapOperand(rule.leftOperand);
-
-  switch (rule.operator) {
-    case "==":
-    case "===":
-    case "!=":
-    case "!==":
-    case ">":
-    case ">=":
-    case "<":
-    case "<=":
-      return `${left} ${rule.operator} ${wrapOperand(rule.rightOperand)}`;
-
-    case "contains":
-      return `String(${left}).includes(${wrapOperand(rule.rightOperand)})`;
-
-    case "startsWith":
-      return `String(${left}).startsWith(${wrapOperand(rule.rightOperand)})`;
-
-    case "endsWith":
-      return `String(${left}).endsWith(${wrapOperand(rule.rightOperand)})`;
-
-    case "isEmpty":
-      return `(${left} === null || ${left} === undefined || ${left} === "")`;
-
-    case "isNotEmpty":
-      return `(${left} !== null && ${left} !== undefined && ${left} !== "")`;
-
-    case "exists":
-      return `(${left} !== null && ${left} !== undefined)`;
-
-    case "doesNotExist":
-      return `(${left} === null || ${left} === undefined)`;
-
-    case "matchesRegex":
-      return `new RegExp(${wrapOperand(rule.rightOperand)}).test(String(${left}))`;
-
-    default: {
-      const _exhaustive: never = rule.operator;
-      return `${left} == ${wrapOperand(rule.rightOperand)}`;
-    }
-  }
-}
-
-/** True when a rule is unfilled — no left operand, or a binary operator with no right operand. */
-function isEmptyRule(rule: ConditionRule): boolean {
-  if (rule.leftOperand.trim() === "") {
-    return true;
-  }
-  if (!isUnaryOperator(rule.operator) && rule.rightOperand.trim() === "") {
-    return true;
-  }
-  return false;
-}
-
-function groupToExpression(group: ConditionGroup): string {
-  if (group.rules.length === 0) {
-    return "true";
-  }
-
-  const joiner = group.logic === "AND" ? " && " : " || ";
-
-  const parts: string[] = [];
-  for (const item of group.rules) {
-    if (isConditionGroup(item)) {
-      const nested = groupToExpression(item);
-      if (nested !== "true") {
-        parts.push(`(${nested})`);
-      }
-    } else if (!isEmptyRule(item)) {
-      parts.push(ruleToExpression(item));
-    }
-  }
-
-  if (parts.length === 0) {
-    return "true";
-  }
-
-  if (parts.length === 1) {
-    return parts[0];
-  }
-
-  return parts.join(joiner);
-}
-
-export function visualConditionToExpression(group: ConditionGroup): string {
-  return groupToExpression(group);
-}
+export { visualConditionToExpression } from "./condition-expression";
 
 // ---------------------------------------------------------------------------
 // Expression -> ConditionGroup parser (reverse of visualConditionToExpression)
